Create open authenticator progress timer indicator
